Add GitHub Actions for new releases
This commit introduces a workflow for handling new releases. It triggers when a release with tag v*.*.* is published. The workflow entails updating the version in package.json, creating a pull request for the version update, and finally publishing the packagei on the registry.

Also, improves .gitignore by ignoring .idea files and modifies the description of the package in package.json.