We need to resubscribe after reconnecting.
This is quite difficult to get right. There is no point attempting to
resubscribe until the reconnection has finished, it will simply fail
immediately. Handle reconnection by pushing onto a new Subject of topics
which need reconnecting.

We don't want to attempt reconnection too often. Buffer failures until
we've seen no new failures for 5s, then reconnect and reattempt the
subscriptions.

This does not leave any mechanism for cancelling a subscription attempt.
This is consistent with the existing behaviour of not attempting to
unsubscribe, but more damaging, as this leads to reconnections.