Handle subscription reconnection better
This is non-trivial. Instead of trying to handle reconnection on the end
of every subscription request, properly track the state of all our
subscriptions. When subscriptions fail, collect up the failures,
reconnect and retry.

This will not resubscribe to topics we are subscribed to; however it
does not refcount subscriptions, so subscribe, subscribe, unsubscribe
will leave us unsubscribed.